Moving middleware to express folder
